Add check for checking existance of previous shortcuts before removing them

+++ lib/jaax.sh
@@ -7,10 +7,9 @@ jaax_dt() {
 	if [ $VERBOSITY -ge $LEV_V ]; then tynnyri "$1-$2"; fi
 	case $1 in
 		clr)
-			if [ -f ~/$KR_DIR_DT/$KR_NAME_LNK ]; then
-			cp ~/$KR_DIR_DT/*.url $KR_DIR_LNK/BU.$HOSTNAME/DESKTOP/; rm ~/$KR_DIR_DT/*.url
-			cp ~/$KR_DIR_DT/*.lnk $KR_DIR_LNK/BU.$HOSTNAME/DESKTOP/; rm ~/$KR_DIR_DT/*.lnk
-			fi;;
+			if [ -f ~/$KR_DIR_DT/$KR_NAME_LNK ]; then cp ~/$KR_DIR_DT/*.lnk $KR_DIR_LNK/BU.$HOSTNAME/DESKTOP/; rm ~/$KR_DIR_DT/*.lnk; fi
+			if [ -f ~/$KR_DIR_DT/$KR_NAME_URL ]; then cp ~/$KR_DIR_DT/*.url $KR_DIR_LNK/BU.$HOSTNAME/DESKTOP/; rm ~/$KR_DIR_DT/*.url; fi
+			;;
 		c) if [ -d $KR_DIR_LNK/COMP/$2 ]; then cp $KR_DIR_LNK/COMP/$2/*.* ~/$KR_DIR_DT/; else virhe JAAX dt $1 $2; fi;;
 		l) if [ -d $KR_DIR_LNK/KTXT/$2 ]; then cp $KR_DIR_LNK/KTXT/$2/*.* ~/$KR_DIR_DT/; else virhe JAAX dt $1 $2; fi;;
 		p) if [ -d $KR_DIR_LNK/PROJ/$2 ]; then cp $KR_DIR_LNK/PROJ/$2/*.* ~/$KR_DIR_DT/; else virhe JAAX dt $1 $2; fi;;
